Add User Password , Add Creator to Video

+++ controllers/videoController.js
@@ -1,6 +1,7 @@
 import routes from "../routes";
 import Video from "../models/Video";
 
+// Home
 export const home = async (req, res) => {
   try {
     const videos = await Video.find({});
@@ -11,6 +12,7 @@ export const home = async (req, res) => {
   }
 };
 
+// Search
 export const search = async (req, res) => {
   const {
     query: { term: searchingBy },
@@ -26,6 +28,7 @@ export const search = async (req, res) => {
   res.render("search", { pageTitle: "Search", searchingBy, videos });
 };
 
+// Video Upload
 export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "Upload" });
 
@@ -38,29 +41,38 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
+// Video Detail
 export const videoDetail = async (req, res) => {
   const {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
   }
 };
 
+// Edit Video
 export const getEditVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -79,12 +91,18 @@ export const postEditVideo = async (req, res) => {
   }
 };
 
+// Video Delete
 export const deleteVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
